+++ src/Listener/ControllerResponseListener.php
@@ -3,7 +3,6 @@
 namespace Torr\Rad\Listener;
 
 use Symfony\Component\EventDispatcher\Attribute\AsEventListener;
-use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpKernel\Event\ViewEvent;
 use Symfony\Component\HttpKernel\KernelEvents;
 use Torr\Rad\Api\ApiResponse;
@@ -27,10 +26,7 @@ final class ControllerResponseListener
 		if ($result instanceof ApiResponse)
 		{
 			$event->setResponse(
-				new JsonResponse(
-					$this->apiResponseNormalizer->normalize($result),
-					$result->statusCode,
-				),
+				$this->apiResponseNormalizer->createResponse($result),
 			);
 		}
 	}